+++ unoxidized/src/synth/params.rs
@@ -7,7 +7,7 @@ impl Synth {
     /**
     Set the master gain
      */
-    pub fn set_gain(&mut self, gain: f64) {
+    pub fn set_gain(&mut self, gain: f32) {
         self.gain = if gain < 0.0 {
             0.0
         } else if gain > 10.0 {
@@ -27,7 +27,7 @@ impl Synth {
     /**
     Get the master gain
      */
-    pub fn get_gain(&self) -> f64 {
+    pub fn get_gain(&self) -> f32 {
         self.gain
     }
